backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
PyJWT==2.8.0
requests==2.31.0
PyYAML==6.0.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
mysql-connector-python==8.2.0